refactor: #52 refactor security files
